backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
requests
xgboost
python-dotenv
